# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DevLearn API — programming language learning platform for senior developers
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /topics:
    get:
      operationId: listTopics
      tags: [topics]
      summary: List topics, optionally filtered by language
      parameters:
        - name: languageId
          in: query
          schema:
            type: integer
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
      responses:
        "200":
          description: List of topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
    post:
      operationId: createTopic
      tags: [topics]
      summary: Create a new topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicInput"
      responses:
        "201":
          description: Created topic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
  /topics/{id}:
    get:
      operationId: getTopic
      tags: [topics]
      summary: Get a single topic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Topic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
        "404":
          description: Not found
...
  /syntax-lessons:
    get:
      operationId: listSyntaxLessons
      tags: [syntax]
      summary: List syntax lessons, optionally filtered by language
      parameters:
        - name: languageId
          in: query
          schema:
            type: integer
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of syntax lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SyntaxLesson"
  /syntax-lessons/{id}:
    get:
      operationId: getSyntaxLesson
      tags: [syntax]
      summary: Get a single syntax lesson with full content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Syntax lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyntaxLesson"
        "404":
          description: Not found
...
components:
  schemas:
...
    Topic:
      type: object
      required: [id, languageId, title, description, difficulty, category]
      properties:
        id:
          type: integer
        languageId:
          type: integer
        languageName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        category:
          type: string
        createdAt:
          type: ["string", "null"]
    TopicInput:
      type: object
      required: [languageId, title, description, difficulty, category]
      properties:
        languageId:
          type: integer
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        category:
          type: string
...
    SyntaxLesson:
      type: object
      required: [id, languageId, title, concept, difficulty]
      properties:
        id:
          type: integer
        languageId:
          type: integer
        languageName:
          type: ["string", "null"]
        title:
          type: string
        concept:
          type: string
        rawSyntax:
          type: ["string", "null"]
        explanation:
          type: ["string", "null"]
        realWorldExample:
          type: ["string", "null"]
        githubProject:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        category:
          type: ["string", "null"]
        orderIndex:
          type: ["integer", "null"]
...