# lib/api-zod/src/generated/types/topicInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DevLearn API — programming language learning platform for senior developers
 * OpenAPI spec version: 0.1.0
 */
import type { TopicInputDifficulty } from './topicInputDifficulty';

export interface TopicInput {
  languageId: number;
  title: string;
  description: string;
  difficulty: TopicInputDifficulty;
  category: string;
}